car_picker/requirements.txt
streamlit>=1.25.0
pandas>=1.5.0
Pillow>=9.0.0
